Separate smart tab setting into separate module for easier tree shaking
